rpk: move acl user to its own package
To keep consistency with the rest of the codebase
